refactor(chat): fold stream-composition tests into stream-route, inline single-use wrapper
- merge stream-composition.test.ts into stream-route.test.ts, reusing its mock scaffold
- fold empty-response persistence check into the existing exhaustion test
- inline prepareMessagesForLLM back into the route (single caller)
- un-export CompactionStreamEvent, trim stale module docs
